requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.0.1
rich==13.9.4
typer==0.15.1
